% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_check.R
\name{col_check}
\alias{col_check}
\title{Check a vector of column names are all contained in a suppled data frame}
\usage{
col_check(df, cols)
}
\arguments{
\item{df}{data frame}

\item{cols}{character vector of column names to be checked whether they exist in \code{df}}
}
\value{
\code{TRUE} if all \code{cols} exist in \code{df}, otherwise an error is thrown
}
\description{
Check a vector of column names are all contained in a suppled data frame
}
\examples{
library(tibble)
adf <- data_frame(A = 1:2, B = 3:4, `C=1` = LETTERS[1:2])
col_check(adf, LETTERS[1:2])
col_check(adf, c("A", "B", "C=1"))
}
\author{
Ty Stanford <tyman@lbtinnovations.com>
}
