% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_c_ik.R
\name{get_c_ik}
\alias{get_c_ik}
\title{Add 2D arrays of [resolver, index] observed counts and [resolver, index] simulated counts}
\usage{
get_c_ik(m_ik, m_st_ik)
}
\arguments{
\item{m_ik}{a [resolver, index] 2D array of counts. Elements are \eqn{\sum_{j=1}^K m_{ijk}} values.}

\item{m_st_ik}{a [resolver, index] 2D array of counts. Elements are \eqn{\sum_{j=1}^K m^*_{ijk}} values.}
}
\value{
a [resolver, index] 2D array of counts. Elements are \eqn{c_{ik}} values.
}
\description{
Add 2D arrays of [resolver, index] observed counts and [resolver, index] simulated counts
}
\examples{
library(tibble)
set.seed(1234)
ABC <- data_frame(
  A=sample(1:3, 20, replace=TRUE),
  B=sample(1:3, 20, replace=TRUE),
  C=sample(c(NA, 1:3), 20, replace=TRUE)
)
m <- get_m_ik(get_m_ijk(ABC, "A", "B", "C"))
p <- get_p_ijk(get_m_ijk(ABC, "A", "B", "C"))
s_st <- get_s_st_ij(ABC, "A", "B", "C")
m_st <- get_m_st_ik(p, s_st)
get_c_ik(m, m_st)
}
\author{
Ty Stanford <tyman@lbtinnovations.com>
}
