% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_m_ijk.R
\name{get_m_ijk}
\alias{get_m_ijk}
\title{Calculate 3D array of [resolver, index, imperfect] counts}
\usage{
get_m_ijk(dat, index, imperfect, resolver)
}
\arguments{
\item{dat}{data frame that contains the index test, imperfect truth and resolver columns}

\item{index}{string of column name in \code{dat} corresponding to the index test}

\item{imperfect}{string of column name in \code{dat} corresponding to the imperfect truth}

\item{resolver}{string of column name in \code{dat} corresponding to the resolver}
}
\value{
a [resolver, index, imperfect] 3D array of counts. Elements are \eqn{m_{ijk}} values. Rows where resolver are \code{NA} are removed.
}
\description{
Calculate 3D array of [resolver, index, imperfect] counts from a dataset
}
\examples{
library(tibble)
set.seed(1234)
ABC <- data_frame(
  A=sample(1:3, 20, replace=TRUE),
  B=sample(1:3, 20, replace=TRUE),
  C=sample(c(NA, 1:3), 20, replace=TRUE)
)
get_m_ijk(ABC, "A", "B", "C")
}
\author{
Ty Stanford <tyman@lbtinnovations.com>
}
