% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_m_ik.R
\name{get_m_ik}
\alias{get_m_ik}
\title{Get 2D array of [resolver, index] counts from 3D array of [resolver, index, imperfect] counts}
\usage{
get_m_ik(m_ijk)
}
\arguments{
\item{m_ijk}{a [resolver, index, imperfect] 3D array of counts}
}
\value{
a [resolver, index] 2D array of counts. Elements are \eqn{\sum_{j=1}^K m_{ijk}} values.
}
\description{
Get 2D array of [resolver, index] counts from 3D array of [resolver, index, imperfect] counts
}
\examples{
library(tibble)
set.seed(1234)
ABC <- data_frame(
  A=sample(1:3, 20, replace=TRUE),
  B=sample(1:3, 20, replace=TRUE),
  C=sample(c(NA, 1:3), 20, replace=TRUE)
)
get_m_ik(get_m_ijk(ABC, "A", "B", "C"))
}
\author{
Ty Stanford <tyman@lbtinnovations.com>
}
