% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_m_st_ik.R
\name{get_m_st_ik}
\alias{get_m_st_ik}
\title{Simulate 2D array of [imperfect, index] counts}
\usage{
get_m_st_ik(p_ijk, s_st_ij)
}
\arguments{
\item{p_ijk}{a [resolver, index, imperfect] 3D array of \eqn{p_{k|ij}}=P(resolver=k | index=i, imperfect=j)}

\item{s_st_ij}{a [imperfect, index] 2D array of counts where resolver value aren't observed (\code{NA}s).}
}
\value{
a [resolver, index] 2D array of simulated counts. Elements are \eqn{\sum_{j=1}^K m^*_{ijk}} values.
}
\description{
Simulate 2D array of [imperfect, index] counts
}
\examples{
library(tibble)
set.seed(1234)
ABC <- data_frame(
  A=sample(1:3, 20, replace=TRUE),
  B=sample(1:3, 20, replace=TRUE),
  C=sample(c(NA, 1:3), 20, replace=TRUE)
)
get_s_st_ij(ABC, "A", "B", "C")
get_p_ijk(get_m_ijk(ABC, "A", "B", "C"))
get_m_st_ik(get_p_ijk(get_m_ijk(ABC, "A", "B", "C")), get_s_st_ij(ABC, "A", "B", "C"))
}
\author{
Ty Stanford <tyman@lbtinnovations.com>
}
