% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_ijk.R
\name{get_p_ijk}
\alias{get_p_ijk}
\title{Calculate 3D array of \eqn{p_{k|ij}} values}
\usage{
get_p_ijk(m_ijk)
}
\arguments{
\item{m_ijk}{a [resolver, index, imperfect] 3D array of \eqn{m_{ijk}} values}
}
\value{
a [resolver, index, imperfect] 3D array of \eqn{p_{k|ij}}=P(resolver=k | index=i, imperfect=j).
}
\description{
Calculate 3D array of \eqn{p_{k|ij}} values.
}
\examples{
library(tibble)
set.seed(1234)
ABC <- data_frame(
  A=sample(1:3, 20, replace=TRUE),
  B=sample(1:3, 20, replace=TRUE),
  C=sample(c(NA, 1:3), 20, replace=TRUE)
)
get_m_ijk(ABC, "A", "B", "C")
get_p_ijk(get_m_ijk(ABC, "A", "B", "C"))
}
\author{
Ty Stanford <tyman@lbtinnovations.com>
}
