% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_s_st_ij.R
\name{get_s_st_ij}
\alias{get_s_st_ij}
\title{Calculate 2D array of [imperfect, index] counts where resolver is \code{NA}}
\usage{
get_s_st_ij(dat, index, imperfect, resolver)
}
\arguments{
\item{dat}{data frame that contains the index test, imperfect truth and resolver columns}

\item{index}{string of column name in \code{dat} corresponding to the index test}

\item{imperfect}{string of column name in \code{dat} corresponding to the imperfect truth}

\item{resolver}{string of column name in \code{dat} corresponding to the resolver}
}
\value{
a [imperfect, index] 2D array of counts. Elements are \eqn{s^*_{ij}} values. Only rows where resolver are \code{NA} are used.
}
\description{
Calculate 2D array of [imperfect, index] counts where resolver is \code{NA} from dataset
}
\examples{
library(tibble)
set.seed(1234)
ABC <- data_frame(
  A=sample(1:3, 20, replace=TRUE),
  B=sample(1:3, 20, replace=TRUE),
  C=sample(c(NA, 1:3), 20, replace=TRUE)
)
get_s_st_ij(ABC, "A", "B", "C")
}
\author{
Ty Stanford <tyman@lbtinnovations.com>
}
