% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_mcrs.R
\name{perform_mcrs}
\alias{perform_mcrs}
\title{Perform mCRS using a one sample-per-row input dataset}
\usage{
perform_mcrs(dat, index, imperfect, resolver, pos, r = 1000,
  seed = 12345678, alpha = 0.05, summ = TRUE)
}
\arguments{
\item{dat}{data frame that contains the index test, imperfect truth and resolver columns (see details)}

\item{index}{string of column name in \code{dat} corresponding to the index test}

\item{imperfect}{string of column name in \code{dat} corresponding to the imperfect truth}

\item{resolver}{string of column name in \code{dat} corresponding to the resolver}

\item{pos}{the levels of \code{index}/\code{imperfect}/\code{resolver} that are considered "positive"}

\item{r}{the number of random simulations of \code{m*_ik} to generate (default=1000)}

\item{seed}{for reporducibility of results (default=12345678)}

\item{alpha}{confidence level (default=0.05)}

\item{summ}{the full \code{r} results are returned when \code{FALSE} or just the summarised values (default, \code{TRUE})}
}
\value{
A tibble with columns \code{param}, \code{cases}, \code{correct}, \code{est}, \code{lo}, \code{up}. The number of rows is determined by the value of \code{summ}
}
\description{
Performs the mCRS analysis
}
\details{
Currently \code{perform_mcrs} requires \code{dat} to contain variables with levels that are consectutive integers starting at \code{1} (i.e. here the outcome levels are labelled 1, 2, 3, 4).
}
\examples{
library(tibble)
data(brenton2018)
# the sens+spec estimates using the imperfect truth
get_sens_spec(table(brenton2018[["S"]], brenton2018[["A"]]), 3:4)
# now check against mCRS (allow ~0.02 sec per repetition)
perform_mcrs(brenton2018, "A", "S", "P", pos = 3:4, r = 10)
}
\author{
Ty Stanford <tyman@lbtinnovations.com>
}
