% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mstar.R
\name{sim_mstar}
\alias{sim_mstar}
\title{Calculate a simulated [resolver, index, imperfect] 3D array}
\usage{
sim_mstar(p_ijk, s_st_ij)
}
\arguments{
\item{p_ijk}{a [resolver, index, imperfect] 3D array of probabilities. Elements are \eqn{p_{k|ij}}=P(resolver|index,imperfect).}

\item{s_st_ij}{a [index, imperfect] 2D array counts requiring a simulated resolver value}
}
\value{
a [resolver, index, imperfect] 3D array of counts. Elements are \eqn{m^*_{k|ij}} values.
}
\description{
Calculate a simulated [resolver, index, imperfect] 3D array
}
\examples{
library(tibble)
set.seed(1234)
AB <- data_frame(A=sample(1:3, 20, replace=TRUE), B=sample(1:3, 20, replace=TRUE))
get_sens_spec(with(AB,table(A,B)), pos = 2:3)
}
\author{
Ty Stanford <tyman@lbtinnovations.com>
}
