% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_dim_col_wise_prop.R
\name{two_dim_col_wise_prop}
\alias{two_dim_col_wise_prop}
\title{Calculate column-wise proportions of counts of a matrix}
\usage{
two_dim_col_wise_prop(x)
}
\arguments{
\item{x}{a square matrix (or object of class "table")}
}
\value{
an object of the same dim as \code{x}.
The elements are the original elements divided by the column sum.
However, if the column sum is 0, the column is replaced by the corresponding column in the equivalent size identity matrix.
}
\description{
Calculate column-wise proportions of counts of a matrix (2D array)
}
\examples{
A <- matrix(c(1,2,1,0,0,0,6,3,3), nrow=3)
two_dim_col_wise_prop(A)
}
\author{
Ty Stanford <tyman@lbtinnovations.com>
}
